requirements.txt
asgiref==3.7.2
autopep8==2.0.2
Django==4.2.4
django-filter==23.2
djangorestframework==3.14.0
Pillow==10.0.0
psycopg2==2.9.7
psycopg2-binary==2.9.7
pycodestyle==2.11.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
